Splite post reset password new password actions
