+++ src/login/app_config/login-rootSaga.js
@@ -4,6 +4,7 @@ import { requestConfig } from "../app_init/init_sagas";
 import { postEmailLink } from "../redux/sagas/resetpassword/postResetPasswordSaga";
 import { useLinkCode } from "../redux/sagas/resetpassword/postVerifyEmailSaga";
 import * as postResetPasswordActions from "../redux/actions/postResetPasswordActions";
+import * as postResetNewPasswordActions from "../redux/actions/postResetNewPasswordActions";
 import loginSagas from "../redux/sagas/rootSaga/loginSagas";
 import { requestPhoneCode } from "../redux/sagas/resetpassword/postExtraSecurityPhoneSaga";
 import { postSetNewPassword } from "../redux/sagas/resetpassword/postSetNewPasswordSaga";
@@ -23,7 +24,7 @@ function* rootSaga() {
     takeLatest(postResetPasswordActions.POST_RESET_PASSWORD, postEmailLink),
     takeLatest(init_actions.GET_JSCONFIG_LOGIN_CONFIG_SUCCESS, useLinkCode),
     takeLatest(postResetPasswordActions.POST_RESET_PASSWORD_EXTRA_SECURITY_PHONE, requestPhoneCode),
-    takeLatest(postResetPasswordActions.POST_RESET_PASSWORD_NEW_PASSWORD, postSetNewPassword),
+    takeLatest(postResetNewPasswordActions.POST_RESET_PASSWORD_NEW_PASSWORD, postSetNewPassword),
   ];
 }
