+++ src/login/redux/actions/postResetNewPasswordActions.js
@@ -1,0 +1,23 @@
+export const POST_RESET_PASSWORD_NEW_PASSWORD = "POST_RESET_PASSWORD_NEW_PASSWORD";
+export const POST_RESET_PASSWORD_NEW_PASSWORD_SUCCESS = "POST_RESET_PASSWORD_NEW_PASSWORD_SUCCESS";
+export const POST_RESET_PASSWORD_NEW_PASSWORD_FAIL = "POST_RESET_PASSWORD_NEW_PASSWORD_FAIL";
+export const POST_RESET_PASSWORD_NEW_PASSWORD_EXTRA_SECURITY_PHONE = "POST_RESET_PASSWORD_NEW_PASSWORD_EXTRA_SECURITY_PHONE";
+export const POST_RESET_PASSWORD_NEW_PASSWORD_EXTRA_SECURITY_PHONE_FAIL = "POST_RESET_PASSWORD_NEW_PASSWORD_EXTRA_SECURITY_PHONE_FAIL";
+
+export function setNewPassword() {
+  return {
+    type: POST_RESET_PASSWORD_NEW_PASSWORD
+  };
+}
+
+export function setNewPasswordFail() {
+  return {
+    type: POST_RESET_PASSWORD_NEW_PASSWORD_FAIL
+  };
+}
+
+export function setNewPasswordExtraSecurityPhoneFail() {
+  return {
+    type: POST_RESET_PASSWORD_NEW_PASSWORD_EXTRA_SECURITY_PHONE_FAIL
+  };
+}
